refactor(backend): add sub-traits and extract task loading
Phase 1 of codebase refactoring:

1. Backend Trait Sub-traits Architecture
   - Create src/backend/traits/ module with focused sub-traits:
     - BackendIdentity: core identity (id, type, plugin association)
     - VersionProvider: version discovery and listing
     - Installer: installation and uninstallation
     - LockfileSupport: lockfile and platform metadata
     - DependencyManager: dependency resolution
     - BinPathProvider: binary paths and execution environment
   - Add comprehensive documentation to Backend trait
   - Add section comments organizing methods by sub-trait category
   - Re-export sub-traits from backend module

2. Task Loading Extraction
   - Create src/config/task_loader.rs module (~500 lines)
   - Extract all task loading functions from config/mod.rs
   - Includes monorepo task discovery and loading
   - Public functions delegate to task_loader module
   - config/mod.rs reduced by ~470 lines

This improves code organization by:
- Splitting the 70+ method Backend trait into logical groups
- Separating task loading concerns from config management
- Providing clear documentation of backend capabilities
